backend/modules/enrollment/hardware/card_reader.py
"""
Card Reader Module for RM365-Toolbox
Provides RFID card reading functionality.
"""

import time
import logging
from typing import Optional

logger = logging.getLogger(__name__)

class CardReaderError(Exception):
    """Raised when card reader operations fail."""
    pass

def read_card_uid(timeout: int = 5) -> str:
    """
    Read a card UID from the connected RFID reader.
    
    Args:
        timeout: Maximum time to wait for a card in seconds
        
    Returns:
        str: The card UID as a hex string
        
    Raises:
        CardReaderError: If no card reader is available or card reading fails
    """
    try:
        # Try to import and use the actual hardware library
        # This would be replaced with actual RFID library imports
        # For now, we'll simulate the behavior
        
        logger.info(f"Attempting to read card UID with {timeout}s timeout")
        
        # Simulate card reading - replace with actual hardware code
        # Example for common RFID libraries:
        # import serial
        # import nfc
        # or other RFID libraries
        
        # For development/testing purposes, we'll raise an error
        # indicating hardware is not available
        raise CardReaderError("RFID hardware not available in development environment")
        
        # Example implementation (commented out):
        """
        import serial
        
        # Open serial connection to RFID reader
        ser = serial.Serial('COM3', 9600, timeout=timeout)  # Adjust port as needed
        
        start_time = time.time()
        while time.time() - start_time < timeout:
            if ser.in_waiting > 0:
                data = ser.readline().decode('ascii').strip()
                if data and len(data) >= 8:  # Minimum UID length
                    ser.close()
                    return data
            time.sleep(0.1)
        
        ser.close()
        raise CardReaderError("No card detected within timeout period")
        """
        
    except ImportError as e:
        logger.warning(f"Card reader library not available: {e}")
        raise CardReaderError("Card reader hardware library not installed")
    except Exception as e:
        logger.error(f"Card reader error: {e}")
        raise CardReaderError(f"Failed to read card: {str(e)}")

def test_card_reader() -> bool:
    """
    Test if the card reader is available and working.
    
    Returns:
        bool: True if card reader is available, False otherwise
    """
    try:
        # This would test the actual hardware connection
        # For now, return False to indicate hardware not available
        return False
        
        # Example test implementation:
        """
        import serial
        ser = serial.Serial('COM3', 9600, timeout=1)
        ser.close()
        return True
        """
        
    except Exception:
        return False

if __name__ == "__main__":
    # Test the card reader
    if test_card_reader():
        print("Card reader is available")
        try:
            uid = read_card_uid(timeout=10)
            print(f"Card UID: {uid}")
        except CardReaderError as e:
            print(f"Error reading card: {e}")
    else:
        print("Card reader is not available")
