backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
reportlab
python-barcode
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt
pandas
pyscard
jinja2
psycopg2-binary
python-multipart
requests
certifi
httpx==0.27.*
